security/ldapauth: move from pkg/ccl/ldapccl
Move the LDAP authentication package out of the CCL tree into
pkg/security/ldapauth. The package is named ldapauth (not ldap)
to avoid colliding with the github.com/go-ldap/ldap/v3 dependency.

Informs #164337

Release note: None
