+++ pkg/BUILD.bazel
@@ -39,7 +39,6 @@ ALL_TESTS = [
     "//pkg/ccl/changefeedccl/tableset:tableset_test",
     "//pkg/ccl/changefeedccl/timers:timers_test",
     "//pkg/ccl/changefeedccl:changefeedccl_test",
-    "//pkg/ccl/ldapccl:ldapccl_test",
     "//pkg/ccl/logictestccl/tests/3node-tenant-multiregion:3node-tenant-multiregion_test",
     "//pkg/ccl/logictestccl/tests/3node-tenant:3node-tenant_test",
     "//pkg/ccl/logictestccl/tests/5node:5node_test",
@@ -377,6 +376,7 @@ ALL_TESTS = [
     "//pkg/security/clientcert:clientcert_test",
     "//pkg/security/dyncert:dyncert_test",
     "//pkg/security/jwtauth:jwtauth_test",
+    "//pkg/security/ldapauth:ldapauth_test",
     "//pkg/security/password:password_test",
     "//pkg/security/provisioning:provisioning_test",
     "//pkg/security/sessionrevival:sessionrevival_test",
@@ -1012,8 +1012,6 @@ GO_TARGETS = [
     "//pkg/ccl/changefeedccl:changefeedccl",
     "//pkg/ccl/changefeedccl:changefeedccl_test",
     "//pkg/ccl/gssapiccl:gssapiccl",
-    "//pkg/ccl/ldapccl:ldapccl",
-    "//pkg/ccl/ldapccl:ldapccl_test",
     "//pkg/ccl/logictestccl/tests/3node-tenant-multiregion:3node-tenant-multiregion_test",
     "//pkg/ccl/logictestccl/tests/3node-tenant:3node-tenant_test",
     "//pkg/ccl/logictestccl/tests/5node:5node_test",
@@ -1958,6 +1956,8 @@ GO_TARGETS = [
     "//pkg/security/jwtauth:jwtauth",
     "//pkg/security/jwtauth:jwtauth_test",
     "//pkg/security/jwthelper:jwthelper",
+    "//pkg/security/ldapauth:ldapauth",
+    "//pkg/security/ldapauth:ldapauth_test",
     "//pkg/security/password:password",
     "//pkg/security/password:password_test",
     "//pkg/security/pprompt:pprompt",
