+++ pkg/ccl/ccl_init.go
@@ -12,7 +12,6 @@ import (
 	_ "github.com/cockroachdb/cockroach/pkg/backup"
 	_ "github.com/cockroachdb/cockroach/pkg/ccl/changefeedccl"
 	_ "github.com/cockroachdb/cockroach/pkg/ccl/gssapiccl"
-	_ "github.com/cockroachdb/cockroach/pkg/ccl/ldapccl"
 	_ "github.com/cockroachdb/cockroach/pkg/ccl/multiregionccl"
 	_ "github.com/cockroachdb/cockroach/pkg/ccl/multitenantccl"
 	_ "github.com/cockroachdb/cockroach/pkg/ccl/oidcccl"
@@ -26,6 +25,7 @@ import (
 	_ "github.com/cockroachdb/cockroach/pkg/crosscluster/physical"
 	_ "github.com/cockroachdb/cockroach/pkg/crosscluster/producer"
 	_ "github.com/cockroachdb/cockroach/pkg/security/jwtauth"
+	_ "github.com/cockroachdb/cockroach/pkg/security/ldapauth"
 )
 
 // TestingEnableEnterprise is a no-op. It was deprecated when the core license
