+++ pkg/security/ldapauth/ldap_util.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package ldapccl
+package ldapauth
 
 import (
 	"context"
